makefile

+++ ft_strlcat.c
@@ -10,11 +10,13 @@
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
 	char		*pdst;
 	const char	*psrc;
-	int			dstlen;
+	size_t			dstlen;
 	int			x;
 
 	x = 0;
