+++ ft_strncmp.c
@@ -10,12 +10,14 @@
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
-	while (i < n && *s1 != "\0" && *s2 != "\0")
+	while (i < n && *s1 != '\0' && *s2 != '\0')
 	{
 		if (s1[i] != s2[i])
 			return (s1[i] - s2[i]);
